chore: move everything into core folder
