+++ packages/client/src/index.ts
@@ -1,9 +1,6 @@
 /* istanbul ignore file */
 import SDK from "./LedgerLivePlatformSDK";
 
-export * from "@ledgerhq/wallet-api-families";
-export * from "@ledgerhq/wallet-api-transports";
-export * from "@ledgerhq/wallet-api-types";
-export * from "./serializers";
+export * from "@ledgerhq/wallet-api-core";
 
 export default SDK;
