+++ packages/core/.eslintrc.js
