+++ packages/core/src/families/algorand/serializer.ts
