+++ packages/core/src/families/bitcoin/serializer.ts
