+++ packages/core/src/families/bitcoin/types.ts
