+++ packages/core/src/families/cosmos/serializer.ts
