+++ packages/core/src/families/stellar/serializer.ts
