+++ packages/core/src/transports/index.ts
@@ -1,3 +1,3 @@
 /* istanbul ignore file */
-// eslint-disable-next-line import/prefer-default-export
+export { default as LedgerPlatformApduTransport } from "./LedgerPlatformApduTransport";
 export { default as WindowMessageTransport } from "./windowMessageTransport";
