+++ packages/core/tsconfig.json
@@ -3,6 +3,7 @@
   "compilerOptions": {
     "declaration": true,
     "outDir": "./dist",
+    "exactOptionalPropertyTypes": false,
     "lib": ["DOM"]
   },
   "include": ["src/**/*", "tests/**/*"]
