--- packages/logger/package.json
@@ -1,21 +1,0 @@
-{
-  "name": "@ledgerhq/wallet-api-logger",
-  "version": "0.1.0",
-  "main": "dist/index.js",
-  "types": "dist/index.d.ts",
-  "license": "MIT",
-  "scripts": {
-    "format:check": "prettier --check \"src\"",
-    "format:fix": "prettier --write \"src\"",
-    "lint": "eslint --cache --ext .ts \"src\"",
-    "lint:fix": "eslint --cache --fix --ext .ts \"src\"",
-    "build": "tsup src/index.ts --dts --format esm,cjs,iife"
-  },
-  "devDependencies": {
-    "@tsconfig/node16-strictest-esm": "^1.0.3",
-    "@types/node": "^18.7.18",
-    "eslint": "^7.32.0",
-    "tsup": "^6.2.3",
-    "typescript": "^4.5.2"
-  }
-}
