--- packages/types/tsconfig.json
@@ -1,9 +1,0 @@
-{
-  "extends": "@tsconfig/node16-strictest-esm/tsconfig.json",
-  "compilerOptions": {
-    "declaration": true,
-    "outDir": "./dist",
-    "exactOptionalPropertyTypes": false
-  },
-  "include": ["src/**/*", "tests/**/*"]
-}
